Refactor: #51 - CameraFrameSelectView를 MFView로 바꾸고 MFViewModel 생성하여 코드 리팩토링.
